Create a POC sql-plugin-api module with SparkShimServiceProvider
Make sure that sql-plugin-api classes are added to dist without using unshimmed-common-from-spark311.txt

Fixes #9138
