+++ aggregator/pom.xml
@@ -70,11 +70,14 @@
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-shade-plugin</artifactId>
                 <configuration>
+                    <shadedArtifactAttached>true</shadedArtifactAttached>
+                    <shadedClassifierName>${spark.version.classifier}</shadedClassifierName>
                     <artifactSet>
                         <excludes>
                             <exclude>org.slf4j:*</exclude>
                             <!-- excludes spark-rapids-jni and jucx, will package them in the dist submodule -->
                             <exclude>com.nvidia:spark-rapids-jni:*</exclude>
+                            <exclude>com.nvidia:rapids-4-spark-sql-plugin-api_${scala.binary.version}:*</exclude>
                             <exclude>org.openucx:jucx:*</exclude>
                         </excludes>
                     </artifactSet>
@@ -90,7 +93,7 @@
                     </relocations>
                     <filters>
                         <filter>
-                            <artifact>com.nvidia:rapids-4-spark-aggregator_2.12</artifact>
+                            <artifact>com.nvidia:rapids-4-spark-aggregator_${scala.binary.version}</artifact>
                             <includes>
                                 <include>META-INF/**</include>
                             </includes>
@@ -131,32 +134,6 @@
                     <execution>
                         <id>default-test</id>
                         <phase>none</phase>
-                    </execution>
-                </executions>
-            </plugin>
-            <plugin>
-                <groupId>org.codehaus.mojo</groupId>
-                <artifactId>build-helper-maven-plugin</artifactId>
-                <executions>
-                    <execution>
-                        <!-- WORKAROUND: default-jar does not contain classes by design.
-                             there seems to be a race between shade and empty default-jar
-                             So attaching here manually because maven complains when we try
-                             to disable default-jar to resolve the conflict.-->
-                        <id>attach-shaded-with-spark-version</id>
-                        <phase>package</phase>
-                        <goals>
-                            <goal>attach-artifact</goal>
-                        </goals>
-                        <configuration>
-                            <artifacts>
-                                <artifact>
-                                    <file>${project.build.directory}/${project.artifactId}-${project.version}.jar</file>
-                                    <type>jar</type>
-                                    <classifier>${spark.version.classifier}</classifier>
-                                </artifact>
-                            </artifacts>
-                        </configuration>
                     </execution>
                 </executions>
             </plugin>
